# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MagicTextResult {
  encodedText: string;
  sampleRate: number;
  frameMs: number;
  lpcOrder: number;
  frameCount: number;
  durationSeconds: number;
  encoding: string;
  maxDurationSeconds: number;
  mode: string;
  estimatedBytes: number;
  millionThirtySecondEstimateGb: number;
  reconstructable: boolean;
  regenerative?: boolean;
  note?: string;
}

export interface ErrorResponse {
  error: string;
}

export type VoiceToMagicTextParams = {
  /**
   * Encoding mode. Photon is a tiny regenerative capsule, ultra is approximate/reconstructable, fingerprint is identification only, standard is detailed LPC.
   */
  mode?: VoiceToMagicTextMode;
};

export type VoiceToMagicTextMode =
  (typeof VoiceToMagicTextMode)[keyof typeof VoiceToMagicTextMode];

export const VoiceToMagicTextMode = {
  standard: "standard",
  ultra: "ultra",
  fingerprint: "fingerprint",
  photon: "photon",
} as const;