# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: voice
    description: Voice encoding operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /voice/magic-text:
    post:
      operationId: voiceToMagicText
      tags: [voice]
      summary: Encode voice audio as compact magic text
      description: Accepts a short audio file as a binary request body and returns an LPC-based compact hexadecimal representation.
      parameters:
        - in: query
          name: mode
          schema:
            type: string
            enum: [standard, ultra, fingerprint, photon]
            default: ultra
          description: Encoding mode. Photon is a tiny regenerative capsule, ultra is approximate/reconstructable, fingerprint is identification only, standard is detailed LPC.
      requestBody:
        required: true
        content:
          audio/wav:
            schema:
              type: string
              format: binary
          audio/mpeg:
            schema:
              type: string
              format: binary
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: Encoded voice representation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MagicTextResult"
        "400":
          description: Missing or invalid audio body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Audio could not be encoded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    MagicTextResult:
      type: object
      properties:
        encodedText:
          type: string
        sampleRate:
          type: integer
        frameMs:
          type: integer
        lpcOrder:
          type: integer
        frameCount:
          type: integer
        durationSeconds:
          type: number
        encoding:
          type: string
        maxDurationSeconds:
          type: integer
        mode:
          type: string
        estimatedBytes:
          type: integer
        millionThirtySecondEstimateGb:
          type: number
        reconstructable:
          type: boolean
        regenerative:
          type: boolean
        note:
          type: string
      required:
        - encodedText
        - sampleRate
        - frameMs
        - lpcOrder
        - frameCount
        - durationSeconds
        - encoding
        - maxDurationSeconds
        - mode
        - estimatedBytes
        - millionThirtySecondEstimateGb
        - reconstructable
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error