# lib/api-zod/src/generated/types/magicTextResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MagicTextResult {
  encodedText: string;
  sampleRate: number;
  frameMs: number;
  lpcOrder: number;
  frameCount: number;
  durationSeconds: number;
  encoding: string;
  maxDurationSeconds: number;
  mode: string;
  estimatedBytes: number;
  millionThirtySecondEstimateGb: number;
  reconstructable: boolean;
  regenerative?: boolean;
  note?: string;
}